+++ lib/currency_formatter.ex
@@ -47,8 +47,64 @@ defmodule CurrencyFormatter do
     |> add_subunit_separator
     |> add_padding
     |> split_units_and_subunits
-    |> handle_cents(format)
+    |> handle_cents(format, opts)
     |> set_symbol(format, opts)
+  end
+
+  @doc """
+  Formats a number to currency wrapped in span tags with classes in a Phoenix.HTML safe way
+  You can directly use this in your phoenix templates.
+
+  ## example
+
+      iex> CurrencyFormatter.html_format(123456, "EUR")
+      [
+        safe: [
+          60,
+          "span",
+          [[32, "class", 61, 34, "currency-formatter-symbol", 34]],
+          62,
+          "€",
+          60,
+          47,
+          "span",
+          62
+        ],
+        safe: [
+          60,
+          "span",
+          [[32, "class", 61, 34, "currency-formatter-amount", 34]],
+          62,
+          "1.234,56",
+          60,
+          47,
+          "span",
+          62
+        ]
+      ]
+  """
+  def html_format(number, currency, opts \\ Keyword.new()) do
+    opts = Keyword.put_new(opts, :html, true)
+    format(number, currency, opts)
+  end
+
+  @doc """
+  Formats a number to currency wrapped in span tags with classes.
+  This will return html as a string without any escaping. When using phoenix, consider using the `format_html/1` function
+
+  ## example
+
+      iex> CurrencyFormatter.raw_html_format(123456, "EUR")
+      ~s[<span class="currency-formatter-symbol">€</span><span class="currency-formatter-amount">1.234,56</span>]
+
+      iex> CurrencyFormatter.raw_html_format(123400, "EUR", keep_decimals: true)
+      ~s[<span class="currency-formatter-symbol">€</span><span class="currency-formatter-amount">1.234,00</span>]
+  """
+  def raw_html_format(number, currency, opts \\ Keyword.new()) do
+    number
+    |> html_format(currency, opts)
+    |> Enum.map(&Phoenix.HTML.safe_to_string/1)
+    |> Enum.join("")
   end
 
   @doc """
@@ -231,9 +287,17 @@ defmodule CurrencyFormatter do
   @spec split_units_and_subunits(binary) :: [binary]
   defp split_units_and_subunits(string), do: String.split(string, ",", parts: 2)
 
-  @spec handle_cents(list, map) :: String.t()
-  defp handle_cents([x, "00"], format), do: set_separators(x, format)
-  defp handle_cents([x, y], format), do: set_separators(x, format) <> format["decimal_mark"] <> y
+  @spec handle_cents(list, map, map) :: String.t()
+  defp handle_cents([x, "00" = y], format, opts) do
+    if Keyword.get(opts, :keep_decimals) do
+      set_separators(x, format) <> format["decimal_mark"] <> y
+    else
+      set_separators(x, format)
+    end
+  end
+
+  defp handle_cents([x, y], format, _opts),
+    do: set_separators(x, format) <> format["decimal_mark"] <> y
 
   @spec set_separators(String.t(), map) :: String.t()
   defp set_separators(string, format) do
@@ -259,15 +323,42 @@ defmodule CurrencyFormatter do
   end
 
   @spec set_symbol(String.t(), map, Keyword.t()) :: String.t()
-  defp set_symbol(number_string, %{"symbol_first" => true} = config, opts),
-    do: get_symbol(config, opts) <> number_string
+  defp set_symbol(number_string, %{"symbol_first" => true} = config, opts) do
+    symbol = get_symbol(config, opts)
+    if Keyword.get(opts, :html) do
+      wrap_in_spans(symbol: symbol, amount: number_string)
+    else
+      symbol <> number_string
+    end
+  end
 
-  defp set_symbol(number_string, config, opts), do: number_string <> get_symbol(config, opts)
+  defp set_symbol(number_string, config, opts) do
+    if Keyword.get(opts, :html) do
+      spans = wrap_in_spans(amount: number_string, symbol: get_symbol(config, opts))
+      Phoenix.HTML.safe_to_string(spans)
+    else
+      number_string <> get_symbol(config, opts)
+    end
+  end
 
-  @spec get_symbol(map) :: String.t()
-  defp get_symbol(config), do: config["symbol"]
-  defp get_symbol(config, disambiguate: true), do: get_disambiguous_symbol(config)
-  defp get_symbol(config, _), do: get_symbol(config)
+  defp wrap_in_spans(symbol: symbol, amount: amount) do
+    [wrap("currency-formatter-symbol", symbol), wrap("currency-formatter-amount", amount)]
+  end
+
+  defp wrap_in_spans(amount: amount, symbol: symbol) do
+    [wrap("currency-formatter-amount", amount), wrap("currency-formatter-symbol", symbol)]
+  end
+
+  defp wrap(class, text) do
+    Phoenix.HTML.Tag.content_tag(:span, text, class: class)
+  end
+
+  @spec get_symbol(map, keyword | nil) :: String.t()
+  defp get_symbol(config, opts \\ Keyword.new()) do
+    if Keyword.get(opts, :disambiguate),
+      do: get_disambiguous_symbol(config),
+      else: config["symbol"]
+  end
 
   @spec get_disambiguous_symbol(map) :: String.t()
   defp get_disambiguous_symbol(%{"disambiguate_symbol" => symbol}), do: symbol